+++ src/Steps/TechnicalOrders/Fields/index.ts
@@ -1,3 +1,4 @@
 //Combine all the files into a single export for ease and to prevent name conflicts
 export * from "./TOCounts";
 export * from "./TOAPMigration";
+export * from "./TOFormat";