+++ CHANGELOG.md
@@ -23,6 +23,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Fixed for any bug fixes.
 - Security in case of vulnerabilities.
 
+## [1.0.5] - 2025-12-18
+
+### Added
+
+- Technical Order Format selection
+
 ## [1.0.4] - 2025-03-xx
 
 ### Added